[IMP] stock: usability of routes configuration
It makes it clear which warehouses, products, product categories
are applied to a certain route (also add inverse many2many field)
in the route form.

The push/procurement rules forms are also a little bit uncluttered
to make a clear distinction between where it is applied and what it creates.