[FIX] models: access error due to prefetch of indirectly referenced records
Complements commit 38e685a82207b9ab8fd3e7578cb431442ed48aa8
in light of commit 62b0d99cfe537e7f7f1e8794974230abbcdbfe44,
to really have the correct effect.

When the prefetching failed due to the presence of
extra records in the cache (for which the access is denied),
the `read` operation was indeed retried. However the
result was not stored in the cache because the cache
already held a FailedValue (automatically added when the
prefetch failed).