[FIX] web: form_view: WidgetButton
Buttons that are not stat buttons couldn't use font-awesome icons, as it
automatically fallbacked on the icons in /web/static/src/img/icons. This was
the case for several templates where the icon thus couldn't be found as it was
looking for a font-awesome icon in the icons folder (e.g. fa-cog.png).

Also cleans some unused less rules defined on the stat buttons in
enterprise_compatibility.less