[FIX] web_kanban, web: kanban title direction
In internet explorer, writing direction can lead to wrong text direction
when combined with rotate.

A fix was already in place for IE up to version 9 but the issue also
happens in next versions which are more difficult to differentiate
(especially the current spartan/edge version).

So this fix use the same css property value for all browser and it doesn't
seem to have an impact (other than correcting the issue in IE).

closes #7955
opw-646430