[FIX][IMP] website_event_track: clean management of track proposal.
Instead of posting a chatter message with the post data + storing a blob on the track
store directly the various data on the track object itself.

Post a message on the event to warn a new track has been created.

Updated the track website view. Instead of storing a blob, rebuild a correct
and editable track view on the website.