[IMP] stock: instead of splitting the pack operation by lots, we allow for multiple lots by pack operation
When scanning multiple serial numbers e.g., the list of pack operations
can become quite long.  In order to avoid that, we make it possible
for pack operations to contain multiple lots.

The implementation is done in the way that the links between moves
and pack operations remains the way it was.  The important difference
is that on action_done, it will check all the reserved quants and their lots at
once first and only afterwards, it will check the lots one by one to check what
it can reserve.

As the details for the lots for the pack operation became a real object, the wizards related
to the pack operation details are not necessary anymore and were replaced
by a form view of the pack operation.