[REV] website_crm,website_sale: revert ce7bac9
The website_crm module doesn't depend on website_sale, so the
salesteam_id field may not exist on website in contactus.

There is clearly a strange overlap with the website salesteam
between website_crm and website_sale, as indicated by the xml_id's.

The previous version didn't crash when the xml_id was not found,
though, while this one would always crash when you try to use
website_crm without website_sale, so I prefer to revert it.

A proper fix would be to set the salesteam at the level of the
website module, but I'm not sure it would make sense considering
that it wouldn't be used at all unless you happen to install one
of those two modules.

In the meantime, I'm reverting ce7bac944a7eecad55237ca1d756cbf6d3d40113.