[FIX] web_kanban: prevent two small issues
when grouping selection fields, the drag and droe was broken (it looked like it
worked, but the id was not saved properly). The problem was that the column id was
not saved in the group.

Another problem was that the creation of new columns was possible when grouping
on non many2one fields