[FIX] base: adapt _serve_attachment for saas-6
Commit 75237918ee15c94750671d5c9c352b4adb9ba786 modified
the logic of _serve_attachment for empty files. After this
patch, empty files are returned with status 200 and a valid
ETag, for proper browser caching.
However the new stored `checksum` field in saas-6 was not
being computed for empty files, resulting in errors when
those files were being served.
This commit updates the logic to restore the caching
behavior for empty files after the forward-port to saas-6.