[FIX] report: only set `--viewport-size` for some reports
viewport-size was add by 63e0d39 (and changed by c9c61e2) to force
wkhtmltopdf to use `col-md-*` bootstrap css classes. As this argument
change the `window` size and not the screen size, this lead to smaller
fonts and overflow issues in reports.
Only change it for reports that need it.

Closes #7882