[FIX] sale: in a SO line, changing uom qty must change uos qty
This is a oversight of the forward port
b5c16cd6039eba174b6cbbb498a720e77061be81
from 8.0

The revision in 8.0 is
cee8f21b6e16d4f91d5715a5724309ff07a5c33e

Not updating the product uos qty on updating the uom qty
is wrong, and it has multiple side-effect.

The SO line margin not being correctly recomputed
when changing the quantity is one of them.

opw-647810