[FIX] base: fix ir_ui_view read file in dev mode
Old code was searching at the same time the xmlid with and without the module
name. But if a file contains a template for his module (without prefix module
name) and a template with the same name for an other template, the node which
are returned canbe the bad template.

Eg: in the same file from an other module than website

<template id="test">
  ...
</template>

<template id="website.test">
  ...
</template>

If you are looking for template website.test, the old code was returning the first
"test" instead of "website.test".

Now we look for the more precise xml_id (website.test), before to fallback on the
implicit name (test = without module name).