[FIX] web: pivot groupby by field from inherits
A related field used in a group by (filter with group_by in context)
which isn't stored on the model, but stored in the related model should
be allowed to be grouped by uppon on a pivot view.

In 8.0 it wasn't the case, but when adding it as a
 <filter context="{'group_by':'field_name'}" />
it was added in the list of fields.

In saas-6 these group by are only available in the search view, but in
some instance (the field is not stored directly on the model) this could
lead to an error which is fixed by this commit.

closes #8349
opw-647401