[FIX] hr_timesheet_sheet: '0' values when adding a new line
When a new line was added, all the others changed their value to 0.
This was because of a property : the sheets propertywas read thanks
to a get('sheets') but as it is an array of object, modifying those
object will modify the sheets (without calling any handler attached
to the 'change:sheets' event). It appeared the form_view set_values
method changed those objects.

A simple solution from a timesheet point of vue was to deep clone the
sheets property before using it.