[FIX] google_calendar: all day event stop date
For an event during all day, during 2 days,
e.g. September 9th and September 10th,
Google expects as start date the first day, and as
stop date the day after the last day.
e.g.
start 2015-09-09
stop 2015-09-11

Even if, afterwise, in the Google Agenda interface, it will
be correctly displayed 2015-09-10 as end date.

This must be applied at the event creation AND at the event
update. It was not the case until this revision.

Besides, we now use the event `stop_date` in Odoo, instead
of the duration, in case the duration has been incorrectly
computed.

opw-648867