[FIX] crm_partner_assign: message translation structure
Backport of 2e7c718 without breaking existing translations.
do not forward port in 9 where proper fix has been made at 2e7c718.

Do not generate the messages by putting bits of sentences together.
Other languages do not have the same structure (and 'have not' was not
translated anyway).

For 8.0 keep the same source for the base error message but create a new one for
the negation as it was wrong in other languages.